requirements.txt
# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-login>=0.6.3

# Database
sqlalchemy>=2.0.23
flask-sqlalchemy>=3.1.1

# ML/AI
transformers>=4.36.0
torch>=2.0.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.2
faiss-cpu>=1.7.4
keybert>=0.8.3

# Utilities
requests>=2.31.0
numpy>=1.26.0
pandas>=2.1.0
python-dotenv>=1.0.0
werkzeug>=3.0.1
joblib>=1.3.2
